requirements.txt
weasyprint>=60.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
tavily-python>=0.5.0
tushare>=1.4.0
arxiv>=2.0.0
# v5.0 新增
requests>=2.31.0
lxml>=5.0.0
feedparser>=6.0.0
openpyxl>=3.1.0
# v5.2 新增
python-dateutil>=2.8.0
